refactor: 비즈니스 로직 검증 부분을 validator로 분리
