+++ src/main/java/com/keunsori/keunsoriserver/domain/reservation/domain/validator/ReservationValidator.java
@@ -2,9 +2,13 @@ package com.keunsori.keunsoriserver.domain.reservation.domain.validator;
 
 import static com.keunsori.keunsoriserver.global.exception.ErrorMessage.ANOTHER_RESERVATION_EXISTS;
 import static com.keunsori.keunsoriserver.global.exception.ErrorMessage.INVALID_RESERVATION_TIME;
+import static com.keunsori.keunsoriserver.global.exception.ErrorMessage.RESERVATION_ALREADY_COMPLETED;
+import static com.keunsori.keunsoriserver.global.exception.ErrorMessage.RESERVATION_NOT_EQUAL_MEMBER;
 
 import org.springframework.stereotype.Component;
 
+import com.keunsori.keunsoriserver.domain.member.domain.Member;
+import com.keunsori.keunsoriserver.domain.reservation.domain.Reservation;
 import com.keunsori.keunsoriserver.domain.reservation.dto.requset.ReservationCreateRequest;
 import com.keunsori.keunsoriserver.domain.reservation.repository.ReservationRepository;
 import com.keunsori.keunsoriserver.global.exception.ReservationException;
@@ -30,6 +34,31 @@ public class ReservationValidator {
 
         if (isThereAnotherReservation) {
             throw new ReservationException(ANOTHER_RESERVATION_EXISTS);
+        }
+    }
+
+    public void validateReservationDeletable(Reservation reservation, Member loggedInMember) {
+        validateReservationNotComplete(reservation);
+        if (loggedInMember.isAdmin()) {
+            return;
+        }
+        validateReservationMember(reservation, loggedInMember);
+    }
+
+    public void validateReservationUpdatable(Reservation reservation, Member loggedInMember) {
+        validateReservationMember(reservation, loggedInMember);
+        validateReservationNotComplete(reservation);
+    }
+
+    private void validateReservationMember(Reservation reservation, Member loggedInMember) {
+        if (!reservation.hasMember(loggedInMember)) {
+            throw new ReservationException(RESERVATION_NOT_EQUAL_MEMBER);
+        }
+    }
+
+    private void validateReservationNotComplete(Reservation reservation) {
+        if (reservation.isComplete()) {
+            throw new ReservationException(RESERVATION_ALREADY_COMPLETED);
         }
     }
