--- src/test/java/com/keunsori/keunsoriserver/fixture/MemberFixture.java
@@ -1,9 +1,0 @@
-package com.keunsori.keunsoriserver.fixture;
-
-import com.keunsori.keunsoriserver.domain.member.domain.Member;
-
-public class MemberFixture {
-
-    public static final Member GENERAL_MEMBER = new Member();
-
-}
