--- src/test/java/com/keunsori/keunsoriserver/fixture/ReservationFixture.java
@@ -1,9 +1,0 @@
-package com.keunsori.keunsoriserver.fixture;
-
-import com.keunsori.keunsoriserver.domain.reservation.domain.Reservation;
-
-public class ReservationFixture {
-
-    public static final Reservation RESERVATION_1 = Reservation.builder().build();
-
-}
